Move GoogleAPI to core
