+++ app/core/endpoints_core.py
@@ -8,6 +8,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core import cruds_core, models_core, schemas_core
 from app.core.config import Settings
+from app.core.google_api.google_api import GoogleAPI
 from app.core.groups.groups_type import GroupType
 from app.dependencies import (
     get_db,
@@ -16,7 +17,6 @@ from app.dependencies import (
     is_user_a_member_of,
 )
 from app.modules.module_list import module_list
-from app.utils.google_api.google_api import GoogleAPI
 from app.utils.tools import is_group_id_valid
 
 router = APIRouter(tags=["Core"])
