--- app/utils/google_api/coredata_google_api.py
@@ -1,17 +1,0 @@
-from datetime import datetime
-
-from app.types.core_data import BaseCoreData
-
-
-class GoogleAPICredentials(BaseCoreData):
-    token: str
-    refresh_token: str
-    token_uri: str
-    client_id: str
-    client_secret: str
-    scopes: list[str]
-    expiry: datetime
-
-
-class GoogleAPIOAuthFlow(BaseCoreData):
-    state: str | None = None
